fixes #505
