+++ views/navbar_content.jade
@@ -10,7 +10,6 @@
     ul.nav.navbar-nav
       li: a(href="/request") Request
       li: a(href='/professional') Professional
-      //li: a(href="/compare") Compare
       li: a(href='/ratings') Ratings
       li: a(href='/faq') FAQ
       li: a(href='/page/1') Blog
