+++ web.js
@@ -229,121 +229,6 @@ app.route('/professional').get(function(req, res, next) {
         */
     });
 });
-app.route('/compare').get(function(req, res, next) {
-    var account_ids = ["all"];
-    //TODO all/percentile works incorrectly due to current advquery getting only one steamid's parsed data per match
-    if (!req.query.compare && req.user) {
-        req.query.compare = req.user.account_id.toString();
-    }
-    if (req.query.compare) {
-        account_ids = account_ids.concat(req.query.compare);
-    }
-    account_ids = account_ids.slice(0, 6);
-    console.log(account_ids);
-    async.mapSeries(account_ids, function(account_id, cb) {
-        var qCopy = JSON.parse(JSON.stringify(req.query));
-        delete qCopy.compare;
-        fillPlayerData(account_id, {
-            info: "compare",
-            query: {
-                select: qCopy,
-                /*
-                js_agg: {
-                    "duration": 1,
-                    "first_blood_time": 1,
-                    "level": 1,
-                    "kills": 1,
-                    "deaths": 1,
-                    "assists": 1,
-                    "last_hits": 1,
-                    "denies": 1,
-                    "hero_damage": 1,
-                    "tower_damage": 1,
-                    "hero_healing": 1,
-                    "gold_per_min": 1,
-                    "xp_per_min": 1,
-                    "kills_per_min": 1,
-                    "deaths_per_min": 1,
-                    "assists_per_min": 1,
-                    "last_hits_per_min": 1,
-                    "hero_damage_per_min": 1,
-                    "tower_damage_per_min": 1,
-                    "hero_healing_per_min": 1
-                }
-                */
-            }
-        }, function(err, player) {
-            //create array of results.aggData for each account_id
-            //compute average for aggregations supporting it
-            //mean or median?
-            for (var key in player.aggData) {
-                /*
-                //mean
-                if (player.aggData[key].sum && player.aggData[key].n) {
-                    player.aggData[key].avg = player.aggData[key].sum / player.aggData[key].n;
-                }
-                */
-                //median
-                var arr = [];
-                for (var value in player.aggData[key].counts) {
-                    for (var i = 0; i < player.aggData[key].counts[value]; i++) {
-                        arr.push(Number(value));
-                    }
-                }
-                arr.sort(function(a, b) {
-                    return a - b;
-                });
-                player.aggData[key].avg = arr[Math.floor(arr.length / 2)];
-            }
-            var result = {
-                account_id: account_id,
-                personaname: player.personaname,
-                aggData: player.aggData
-            };
-            cb(err, result);
-        });
-    }, function(err, results) {
-        if (err) {
-            return next(err);
-        }
-        //compute percentile for each stat
-        //for each stat average in each player's aggdata, iterate through all's stat counts and determine whether this average is gt/lt key, then add count to appropriate bucket. percentile is gt/(gt+lt)
-        results.forEach(function(r, i) {
-            for (var key in results[i].aggData) {
-                var avg = results[i].aggData[key].avg;
-                var allCounts = results[0].aggData[key].counts;
-                var gt = 0;
-                var lt = 0;
-                if (avg) {
-                    for (var value in allCounts) {
-                        var valueCount = allCounts[value];
-                        if (avg >= Number(value)) {
-                            gt += valueCount;
-                        }
-                        else {
-                            lt += valueCount;
-                        }
-                    }
-                    results[i].aggData[key].percentile = gt / (gt + lt);
-                }
-            }
-        });
-        res.render("compare", {
-            //TODO list won't include self
-            //TODO list should include cached player teammates and players manually added via tag
-            all_teammate_list: (typeof req.query.compare === "object" ? req.query.compare : [req.query.compare]).map(function(c) {
-                return {
-                    account_id: c,
-                    personaname: c
-                };
-            }),
-            data: results,
-            //TODO query object needs to be processed (convert string to number, etc.)
-            q: req.query,
-            compare: true
-        });
-    });
-});
 app.use('/matches', require('./routes/matches'));
 app.use('/players', require('./routes/players'));
 app.use('/api', require('./routes/api'));
